Change insomnia info, night hours to 0 - 8, added compare row, link to chart. Fixed current date

+++ user_charts.php
@@ -48,7 +48,7 @@ class OnlineHistoryCharts extends OnlineHistory{
 	function get_activity_by_user($users){
 		
 		$users_string = implode(",", $users);
-		$count_query = "SELECT user_id, EXTRACT(hour FROM status) AS hours, COUNT (EXTRACT(hour FROM status)) / 12 AS count 
+		$count_query = "SELECT user_id, EXTRACT(hour FROM status) AS hours, COUNT (EXTRACT(hour FROM status))::float / 12 AS count 
 				FROM user_online 
 				WHERE user_id IN ({$users_string}) 
 				GROUP BY hours, user_id 
@@ -76,7 +76,7 @@ class OnlineHistoryCharts extends OnlineHistory{
 	        $my_date_end = $this->get_correct_date_interval($my_date)[1];
 		
 		$users_string = implode(",", $users);
-		$count_query = "SELECT user_id, status::timestamp::date AS day, COUNT (EXTRACT(hour FROM status)) / 12 AS count 
+		$count_query = "SELECT user_id, status::timestamp::date AS day, COUNT (EXTRACT(hour FROM status))::float / 12 AS count 
                                 FROM user_online
                                 WHERE user_id IN ({$users_string}) AND status between '{$my_date_start}' and '{$my_date_end} 23:59:59' 
                                 GROUP BY user_id, day
